env系のnorm対応
