+++ srcs/executor/pipe_exec_cmd.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   pipe_ut.c                                          :+:      :+:    :+:   */
+/*   pipe_exec_cmd.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: yabukirento <yabukirento@student.42.fr>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/04/08 13:09:57 by yabukirento       #+#    #+#             */
-/*   Updated: 2025/04/08 13:11:25 by yabukirento      ###   ########.fr       */
+/*   Created: 2025/04/08 13:18:06 by yabukirento       #+#    #+#             */
+/*   Updated: 2025/04/08 13:18:07 by yabukirento      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
