+++ srcs/parser/env.c
@@ -3,14 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   env.c                                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: myokono <myokono@student.42.fr>            +#+  +:+       +#+        */
+/*   By: yabukirento <yabukirento@student.42.fr>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/06 21:24:50 by myokono           #+#    #+#             */
-/*   Updated: 2025/04/08 13:07:24 by myokono          ###   ########.fr       */
+/*   Updated: 2025/04/08 13:25:14 by yabukirento      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../includes/minishell.h"
+#include "minishell.h"
 
 static char	*get_last_command_path(t_shell *shell)
 {
@@ -22,20 +22,15 @@ static char	*get_last_command_path(t_shell *shell)
 	return ("/bin/bash");
 }
 
-int	expand_env_var(char *input, int *i, char **result, t_shell *shell)
+static int	handle_special(char *input, int *i, char **result, t_shell *shell)
 {
-	int		start;
-	char	*var_name;
-	char	*var_value;
-
-	(*i)++;
 	if (input[*i] == '?')
 	{
-		var_value = ft_itoa(shell->exit_status);
-		*result = ft_strjoin(*result, var_value);
-		free(var_value);
+		input = ft_itoa(shell->exit_status);
+		*result = ft_strjoin(*result, input);
+		free(input);
 		(*i)++;
-		return (0);
+		return (1);
 	}
 	if (input[*i] == '\'' && input[*i - 1] == '$')
 	{
@@ -50,8 +45,20 @@ int	expand_env_var(char *input, int *i, char **result, t_shell *shell)
 	if (input[*i] == '\'' || input[*i] == '\"')
 	{
 		*result = ft_strjoin(*result, ft_strdup("$"));
-		return (0);
+		return (1);
 	}
+	return (0);
+}
+
+int	expand_env_var(char *input, int *i, char **result, t_shell *shell)
+{
+	int		start;
+	char	*var_name;
+	char	*var_value;
+
+	(*i)++;
+	if (handle_special(input, i, result, shell))
+		return (0);
 	start = *i;
 	while (input[*i] && (ft_isalnum(input[*i]) || input[*i] == '_'))
 		(*i)++;
@@ -65,9 +72,7 @@ int	expand_env_var(char *input, int *i, char **result, t_shell *shell)
 	if (ft_strcmp(var_name, "_") == 0 && !var_value)
 		var_value = get_last_command_path(shell);
 	if (var_value)
-	{
 		*result = ft_strjoin(*result, ft_strdup(var_value));
-	}
 	free(var_name);
 	return (SUCCESS);
 }
