+++ srcs/utils/env_utils.c
@@ -3,44 +3,29 @@
 /*                                                        :::      ::::::::   */
 /*   env_utils.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: myokono <myokono@student.42tokyo.jp>       +#+  +:+       +#+        */
+/*   By: yabukirento <yabukirento@student.42.fr>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/27 13:03:37 by myokono           #+#    #+#             */
-/*   Updated: 2025/04/08 11:24:28 by myokono          ###   ########.fr       */
+/*   Updated: 2025/04/08 13:27:59 by yabukirento      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/minishell.h"
 
-void	update_env_array(t_shell *shell)
+t_env	*create_env_node(char *key, char *value)
 {
-	if (shell->env_array)
-		free_array(shell->env_array);
-	shell->env_array = env_list_to_array(shell->env_list);
-}
+	t_env	*new_node;
 
-t_env	*get_env_node(t_env *env_list, char *key)
-{
-	t_env	*current;
-
-	current = env_list;
-	while (current)
+	new_node = malloc(sizeof(t_env));
+	if (!new_node)
 	{
-		if (ft_strcmp(current->key, key) == 0)
-			return (current);
-		current = current->next;
-	}
-	return (NULL);
-}
-
-char	*get_env_value(t_env *env_list, char *key)
-{
-	t_env	*node;
-
-	node = get_env_node(env_list, key);
-	if (!node)
+		system_error("malloc");
 		return (NULL);
-	return (node->value);
+	}
+	new_node->key = ft_strdup(key);
+	new_node->value = ft_strdup(value);
+	new_node->next = NULL;
+	return (new_node);
 }
 
 void	add_env_node(t_env **env_list, char *key, char *value)
@@ -56,15 +41,9 @@ void	add_env_node(t_env **env_list, char *key, char *value)
 		existing_node->value = ft_strdup(value);
 		return ;
 	}
-	new_node = malloc(sizeof(t_env));
+	new_node = create_env_node(key, value);
 	if (!new_node)
-	{
-		system_error("malloc");
 		return ;
-	}
-	new_node->key = ft_strdup(key);
-	new_node->value = ft_strdup(value);
-	new_node->next = NULL;
 	if (!*env_list)
 	{
 		*env_list = new_node;
@@ -76,10 +55,9 @@ void	add_env_node(t_env **env_list, char *key, char *value)
 	current->next = new_node;
 }
 
-void	remove_env_node(t_env **env_list, char *key)
+void	remove_head_env_node(t_env **env_list, char *key)
 {
 	t_env	*current;
-	t_env	*prev;
 
 	if (!*env_list)
 		return ;
@@ -90,6 +68,19 @@ void	remove_env_node(t_env **env_list, char *key)
 		free(current->key);
 		free(current->value);
 		free(current);
+	}
+}
+
+void	remove_env_node(t_env **env_list, char *key)
+{
+	t_env	*current;
+	t_env	*prev;
+
+	if (!*env_list)
+		return ;
+	if (ft_strcmp((*env_list)->key, key) == 0)
+	{
+		remove_head_env_node(env_list, key);
 		return ;
 	}
 	prev = *env_list;
